Add photo list model and view
The model is a QAbstractListModel that is fed an EventItem and uses that
item's list of PhotoItems. Currently only the path as read from the
database is used in the DisplayRole, which is probably not the proper
role to use. In any case it will be used later for generating
thumbnails.

The view is written in QML. The qml source is packaged as a resource at
compile time.