Sort the tree as Shotwell does
Years, months and events are in decending order in the tree, but within
an event the photos are shown in ascending order by exposure time.

Since there's now a guaranteed sorting, EventItem no longer needs
startTime. Add getEndTime() will we're at it; will be used to generate a
pretty event name later.