Replace PhotoEventListProxyModel with a better proxy
The new FilterFlattenProxyModel takes a template argument which is the
type of element from the event model to show. This can be used (with
separate instances) to provide a list of events and a complete list of
photos.

It's somewhat inspired by KDescendantsPRoxyModel in that it keeps a
complete mapping using QPersistentModelIndex as keys. Expected to be
slow.