Select a photo to show it fullscreen
The user can select a photo by either double clicking or pressing
Enter.

The full screen view is not of a single photo, but just another view
onto the model. I try to hide this my making it not interactive and by
not having any animations. However, users (me included) expect to be
able to browse photos by at least the arrow keys, so let them.

The full image is loaded asynchronously. Properly handle keyboard focus
between the view.