Add temporary views
Temporary in the database connection sense. For now I want to make
Shotwin not alter the database in any way.

These views will be useful for the photo model, and will simplify the
query for the event tree model. I did not know about min() and max()
before. These views will necessitate changes in the Item classes used in
existing models, hopefully for the better.