Rewrite proxy insertion logic
Now handles "arbitrary" depth. The data on which to group by are now
QVariants so there is no longer a need for a distinction between those
and what is used as the DisplayRole.

However, the user *may* supply a source data callback to supply extra
information (also as QVariants) that can be retrieved as the role
SourceDataRole. This can be useful for implementing responses to clicks
etc, instead of infering the actual data from the display data.