Make use of new models
Instead of the EventTreeModel, use the new flat EventModel and
TreeProxyModel. Add a DateFilteredEventModel to QML with a filtered list
of events.

Accomodate the slots in the Shotwin object that handle signals from the
views accordingly.

I am very pleased that this required no changes outside shotwin.h or
shotwin.cpp! Granted, the new slots don't look as pretty.