Support the idea of "alternate display data"
When using the TreeProxyModel and grouping items based on the original
data it might be desired to modify the shown data for the actual node
corresponding to the source model. For example, when grouping a list of
file paths according to the hierarchy the original data might be the
full path but while in the tree only the file name needs to be shown.

TreeProxyModel now accepts an optional callback to modify the data
returned by data() for the display role. The callback receives the
original data as its only argument, and is expected to return the
modified data the user wants show.